Third Update.
Configured a cinemachine camera. Overhauled the enemy AI to detect colliders like the player and wall using raycasting. Implemented a collectible to pick up. Added the Main Menu and hooked it up so that the player  navigate between scenes using the pause or game over screens. Added sound effects to the enemy so that it plays an audio clip every time it attacks the player. Added an audio clip that loops in the background to create ambience.